Refactor v2 validator type inference
